chore: basic password feedback
